Move manifest fetcher into a new script
